Move BNC helpers into BNC adapter
